+++ src/app/public/contact/contact.component.ts
@@ -1,7 +1,6 @@
 import {Component, OnInit} from '@angular/core';
 import {TranslatePipe} from '@ngx-translate/core';
 import {FormBuilder, FormGroup, ReactiveFormsModule, Validators} from '@angular/forms';
-import {CommonModule} from '@angular/common';
 
 @Component({
   selector: 'app-contact',
@@ -9,7 +8,6 @@ import {CommonModule} from '@angular/common';
   imports: [
     TranslatePipe,
     ReactiveFormsModule,
-    CommonModule
   ],
   templateUrl: './contact.component.html',
   styleUrl: './contact.component.css'
@@ -21,7 +19,7 @@ export class ContactComponent implements OnInit {
   isSubmitted = false;
   isLoading = false;
 
-  constructor(private fb: FormBuilder) {
+  constructor(private readonly fb: FormBuilder) {
     this.contactForm = this.fb.group({
       name: ['', Validators.required],
       email: ['', [Validators.required, Validators.email]],
@@ -32,31 +30,32 @@ export class ContactComponent implements OnInit {
 
   ngOnInit(): void {
     this.contact = {
-      name: "",
       address: "Av. B 200 Lima",
-      email: "cbaciliod@gmail.com",
+      email: "bacsystem.sac@gmail.com",
       phone: "+51 907 892 868",
     };
   }
 
-  onSubmit() {
+  onSubmit(): void {
     if (this.contactForm.valid) {
       this.isLoading = true;
-      // Simulate sending
       setTimeout(() => {
         this.isSubmitted = true;
         this.isLoading = false;
-        console.log('Form submitted:', this.contactForm.value);
       }, 2000);
     } else {
       this.contactForm.markAllAsTouched();
     }
   }
+
+  resetForm(): void {
+    this.isSubmitted = false;
+    this.contactForm.reset();
+  }
 }
 
 
 interface Contact {
-  name: string;
   email: string;
   phone: string;
   address: string;