+++ src/app/public/footer/footer.component.spec.ts
@@ -1,4 +1,6 @@
 import { ComponentFixture, TestBed } from '@angular/core/testing';
+import { provideRouter } from '@angular/router';
+import { TranslateModule } from '@ngx-translate/core';
 
 import { FooterComponent } from './footer.component';
 
@@ -8,7 +10,8 @@ describe('FooterComponent', () => {
 
   beforeEach(async () => {
     await TestBed.configureTestingModule({
-      imports: [FooterComponent]
+      imports: [FooterComponent, TranslateModule.forRoot({})],
+      providers: [provideRouter([])],
     })
     .compileComponents();